further modularize iad.c; continue header cleanup
