+++ src/iad/iad.h
@@ -1,10 +1,4 @@
 #ifndef IAD_H
 #define IAD_H
 
-// Function declarations
-void perform_cleanup(void);
-void handle_sigint(int sig);
-
-// Constants (example)
-
 #endif // IAD_H
